SF-1143: Create CircleCI release workflow (#2)
* Consolidate scripts to top level
* Include each package's real version in their respective package.json files
* Add changelogs for all packages
* Move contributing doc to top
* Pull all devDependencies to the top
* Pull codecov.yml to the root of the repo

The CircleCI workflow includes three high-level steps: 1) build, 2) test, and 3) release.

Build

The build step builds all packages. flux-capacitor, core and structure
build in sequence, then the remaining packages build in parallel. The
result is cached and collected in a final build step to be used in later
jobs.

Test

The packages are all tested in parallel.

Release

Each package releases in three steps: 1) bump its dependencies, 2)
update its version number and tag the resulting commit, and 3) publish
it to npm.

The first step updates the version number of a dependency if that
dependency had a release. For example, a release in flux-capacitor will
cause core to bump its dependency version. The resulting release type
for the second component will be the higher of the two release types.

core receives a bump from flux-capacitor. structure receives a bump from
core. All components receive a bump from core and structure (if a
component does not depend on structure, the structure bump is ignored).

Like the build jobs, flux-capacitor, core and structure release in
sequence, and the components release in parallel.

All these commits are made on a temporary release branch, which is later
merged in to master on successful completion.

In addition, there is a job that runs only for non-master branches that
ensures there is no release in progress on master. This will be used as
a check on GitHub to prevent a pull request from being merged in when
there is a release in progress.