Merge in the release branch with a nicer message
Also do a fast-forward merge if there was no release. It is assumed that
if there was no CDN release, no package release was made, which means
that the only commit on the branch is the docs commit. This is merged in
as a fast-forward merge to avoid cluttering the history.