SF-970: Remove unnecessary update on mount for sayt component (#96)
Being active by default and then becoming inactive on mount may have
been necessary when `sayt` had used `if` instead of `show` to allow the
inner components to be instantiated. It is no longer necessary because
`sayt` uses `show`, which allows the inner components to instantiate
while still be hidden from view.